MQE: Tidy up pool size handling
Primarily from feedback on https://github.com/grafana/mimir/pull/10261

Rather than allowing an arbitrary maxSize in pools, require them to
be a power of two since that is what happens internally anyway.

We keep the checks and protections around requiring it to be a power of
two.
