+++ pkg/streamingpromql/types/pool_test.go
@@ -1,0 +1,16 @@
+// SPDX-License-Identifier: AGPL-3.0-only
+
+package types
+
+import (
+	"testing"
+
+	"github.com/grafana/mimir/pkg/util/pool"
+	"github.com/stretchr/testify/require"
+)
+
+func TestMaxExpectedSeriesPerResultConstantIsPowerOfTwo(t *testing.T) {
+	// Although not strictly required (as the code should handle maxExpectedSeriesPerResult not being a power of two correctly),
+	// it is best that we keep it as one for now.
+	require.True(t, pool.IsPowerOfTwo(maxExpectedSeriesPerResult), "maxExpectedSeriesPerResult must be a power of two")
+}
